pyproject.toml
[project]
name = "playground-pipeline-docker"
version = "0.1.0"
description = "Data pipeline using PySpark, Iceberg, and PostgreSQL"
requires-python = ">=3.9"
dependencies = [
    "pyspark==3.5.0",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "pyiceberg==0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.0",
    "pre-commit>=3.0",
    "pylint>=3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pipeline"]

[tool.black]
line-length = 79
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
